fixed command line option types for optuna optimization

+++ src/config/analysis_config.py
@@ -140,14 +140,14 @@ class AnalysisConfig(Config):
             "-cB",
             "--coeff_max_bond",
             default=1,
-            type=int,
+            type=float,
             help="Coefficient for term maximizing number of bonds",
         )
         energy_parser.add_argument(
             "-cL",
             "--coeff_stem_len",
             default=10,
-            type=int,
+            type=float,
             help="Coefficient for term penalizing short stems",
         )
         energy_parser.add_argument(
@@ -161,7 +161,7 @@ class AnalysisConfig(Config):
             "-mu",
             "--target_stem_length",
             default=-1,
-            type=float,
+            type=int,
             help="Value defining one-body energy penalty. Default: -1 (maximum possible stem)",
         )
